+++ src/collective/dms/mailcontent/browser/settings.py
@@ -1,20 +1,35 @@
+from collective.dms.mailcontent import _
 from DateTime import DateTime
-from five import grok
-from Products.CMFCore.interfaces import ISiteRoot
-from zope.interface import Interface
-from zope import schema
-from plone.z3cform import layout
-from plone.app.registry.browser.controlpanel import RegistryEditForm
 from plone.app.registry.browser.controlpanel import ControlPanelFormWrapper
-from Products.PageTemplates.Expressions import getEngine
+from plone.app.registry.browser.controlpanel import RegistryEditForm
+from plone.supermodel import model
+from plone.z3cform import layout
 from Products.CMFCore.Expression import Expression
-from .. import _
+from Products.PageTemplates.Expressions import getEngine
+from zope import schema
+from zope.publisher.browser import BrowserPage
+from zope.schema.vocabulary import SimpleTerm
+from zope.schema.vocabulary import SimpleVocabulary
 
 
-class IDmsMailConfig(Interface):
+im_irn_conditions = SimpleVocabulary([SimpleTerm(u"show", title=_("Always show")),
+                                      SimpleTerm(u"hide", title=_("Always hide"))])
+
+om_irn_conditions = SimpleVocabulary([SimpleTerm(u"show", title=_("Always show")),
+                                      SimpleTerm(u"hide", title=_("Always hide")),
+                                      SimpleTerm(u"reply", title=_("Show only when replying"))])
+
+
+class IDmsMailConfig(model.Schema):
     """
     Configuration of dms mail
     """
+
+    model.fieldset(
+        'incomingmail',
+        label=_(u"Incoming mail"),
+        fields=['incomingmail_number', 'incomingmail_talexpression']
+    )
 
     incomingmail_number = schema.Int(
         title=_(u'Number of next incoming mail'),
@@ -22,7 +37,14 @@ class IDmsMailConfig(Interface):
 
     incomingmail_talexpression = schema.TextLine(
         title=_(u"Incoming mail internal reference default value expression"),
-        description=_(u"Tal expression where you can use portal, number as variable")
+        description=_(u"Tal expression where you can use portal, number, context, request, date as variable")
+    )
+
+    model.fieldset(
+        'outgoingmail',
+        label=_(u"Outgoing mail"),
+        fields=['outgoingmail_number', 'outgoingmail_talexpression', 'outgoingmail_edit_irn',
+                'outgoingmail_increment_number', 'outgoingmail_today_mail_date']
     )
 
     outgoingmail_number = schema.Int(
@@ -31,7 +53,19 @@ class IDmsMailConfig(Interface):
 
     outgoingmail_talexpression = schema.TextLine(
         title=_(u"Outgoing mail internal reference default value expression"),
-        description=_(u"Tal expression where you can use portal, number as variable")
+        description=_(u"Tal expression where you can use portal, number, context, request, date as variable")
+    )
+
+    outgoingmail_edit_irn = schema.Choice(
+        title=_(u"Internal reference number edition field"),
+        vocabulary=om_irn_conditions,
+        default='show'
+    )
+
+    outgoingmail_increment_number = schema.Bool(
+        title=_(u"Increment number used in internal reference number"),
+        description=_(u"If False, number will not be incremented only if user can edit internal reference number"),
+        default=True
     )
 
     outgoingmail_today_mail_date = schema.Bool(
@@ -49,20 +83,17 @@ class SettingsEditForm(RegistryEditForm):
     label = _(u"Dms Mail settings")
 
 
-class SettingsView(grok.View):
+class SettingsView(BrowserPage):
     """
     View which wrap the settings form using ControlPanelFormWrapper to a HTML boilerplate frame.
     """
-    grok.name("dmsmailcontent-settings")
-    grok.context(ISiteRoot)
-    grok.require('plone.app.controlpanel.Site')
 
-    def render(self):
+    def __call__(self):
         view_factor = layout.wrap_form(SettingsEditForm, ControlPanelFormWrapper)
         view = view_factor(self.context, self.request)
         return view()
 
-    def evaluateTalExpression(self, expression, portal, number, **kwargs):
+    def evaluateTalExpression(self, expression, context, request, portal, number, **kwargs):
         """
             evaluate the expression, considering portal and number in context
         """
@@ -70,6 +101,8 @@ class SettingsView(grok.View):
         data = {
             'tool': self,
             'number': str(number),
+            'context': context,
+            'request': request,
             'portal': portal,
             'date': DateTime(),
         }